+++ backend/security/src/main/kotlin/io/tolgee/security/ratelimit/RateLimitInterceptor.kt
@@ -26,6 +26,7 @@ import org.springframework.web.method.HandlerMethod
 import org.springframework.web.servlet.HandlerInterceptor
 import org.springframework.web.servlet.HandlerMapping
 import java.time.Duration
+import javax.servlet.DispatcherType
 import javax.servlet.http.HttpServletRequest
 import javax.servlet.http.HttpServletResponse
 
@@ -36,7 +37,7 @@ class RateLimitInterceptor(
   private val rateLimitService: RateLimitService,
 ) : HandlerInterceptor, Ordered {
   override fun preHandle(request: HttpServletRequest, response: HttpServletResponse, handler: Any): Boolean {
-    if (handler !is HandlerMethod) {
+    if (handler !is HandlerMethod || DispatcherType.ASYNC == request.dispatcherType) {
       return super.preHandle(request, response, handler)
     }